- refactoring file system view by extracting components and services to separate files - extracted components: FileRowView, FolderRowView, PDRGenerator
